ARM: AT91: migrate sama5d3-xplained to DT
We have one other DT-enabled SAMA5D3 board already:
microchip-ksz8477-evb, which is basically a sama5d3 xplained in
different form factor and with a switch. Mimic what we do there to get
rid of the sama5d3 board code.

Unlike the microchip-ksz9477-evb, the barebox support for this board
includes NAND, which is not yet up to date with respect to DT:

  - SMC for NAND configuration has no DT driver. Instead the board
    driver create this device
  - We don't support the newer NAND DT bindings (EBI). The device trees
    fix this up, so barebox sees the old bindings with the upstream
    partition layout. In future, this could be fixed.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210922065000.20970-6-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>