kbuild: dtc: Allow adding device tree fragments via config
This introduces a config variable that allows adding additional fragments
to the Barebox device tree(s).

Example uses are adjusting the flash partition layout, adding barebox
state variables, or adding an I2C device.  These can be now be done with
build configuration only, without needing to patch the existing dts
files in the Barebox source.

The advantage is greater when an external build system, such as Yocto or
Buildroot, is being used to build Barebox.  The build system can drop in
a dts fragment to partition flash and build from unaltered Barebox
source.  This avoids the need for cumbersome maintenance of patch files
to modify Barebox's source for each flash partition layout.

Preprocessing the dts file gains another layer, where a generated dts
source consisting of an include directive for the original dts source is
followed by more includes for each fragment.  This is piped to the
existing preprocessor call on stdin to avoid another temporary file.
cpp/dtc will correctly identify errors in the source files they occur
in.  The -MT option is used so the cpp auto-dependencies reference the
original dts source and not the generated code passed on stdin.

A preprocessor macro named after the base dts file, e.g. foo-bar.dts
will define foo_bar_dts, will be defined so that the fragments can
possibly operate differently based on which image's dts is being built.

Signed-off-by: Trent Piepho <trent.piepho@igorinstitute.com>
Link: https://lore.barebox.org/20210922181336.1350460-1-trent.piepho@igorinstitute.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>