ddr: imx8m: Rework ddrphy code and fix DDR-1067 on iMX8MM
This code could not decide if it was using speeds in Hz, or MHz, or just
constants that identify specific speeds.

Rework it to use symbolic constants for all speeds.  Use arrays
for configurations values for all cases.  The DDR DRAM speed constants
will be the array index, so the linear search for the configuration is
eliminated.

For iMX8MM type SoCs using the integr PLL the speed DDR-1067 with a 266⅔
MHz clock will now work.

The iM8MQ type SSCG PLL SoCs would previously silently program the PLL
will zero values if a non-supported DDR rate was used.  Now they will
generate an error.

Note that some PLL tables have entries for speeds that the main entry
point for the code does not support, so they can't actually ever be
used.  This is not fixed.

Signed-off-by: Trent Piepho <tpiepho@gmail.com>
Tested-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210924031446.1515172-1-tpiepho@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>