nandtest: Add more fields when bad byte found
This can help to determine what is causing the problem.  E.g., is the
bad byte always in the same eraseblock?  Is it always 513 bytes into a
page with a hardware ECC chunk size of 512 bytes?

Example before and after (error at absolute address 123456 with 128kB
eraseblock and 2kB page sizes):

Byte 0xe240 is 12 should be 34
Block 0x1 byte 0xe240 (page 0x1c offset 0x240) is 0x12 should be 0x34

Signed-off-by: Trent Piepho <trent.piepho@igorinstitute.com>
Link: https://lore.barebox.org/20211011190725.856597-1-trent.piepho@igorinstitute.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>