regulator: allow use of dummy regulator
The idea of devicetree property which allows use of dummy regulator is
not new but has not been implemented until now.
This implementation uses barebox specific devicetree property
"barebox,allow-dummy-supply" to allow switching to a dummy power
regulator in cases where proper regulator driver is not available. This
property can be set for regulator or for PMIC regulators nodes, which
then allow use of dummy regulator for all its child nodes. Basically
just catch the regulators ensure_probed error, if
"barebox,allow-dummy-supply" property is set and return dummy regulator.

Signed-off-by: Andrej Picej <andrej.picej@norik.com>
Link: https://lore.barebox.org/20211119095429.1905473-5-andrej.picej@norik.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>