mci: bcm2835: Don't send MMC_CMD_GO_IDLE_STATE at the end of .init()
The core sends this command anyhow shortly after calling .init(). So
doing it in bcm2835_mci_reset() is only overhead if it succeeds and
if there is a problem it is noticed a tad earlier which has little
benefit compared to the timeout in bcm2835_mci_wait_command_done().

So simplify and drop the transfer.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Link: https://lore.barebox.org/20211130203026.360693-1-u.kleine-koenig@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>