ARM64: qemu-virt64: disable MMU
We have CFI flash at address 0 till start of RAM. Until we have a scheme
to remap the flash, so it's usable at that address, just disable the MMU
in the defconfig. Currently the defconfig prints a couple of error
messages on every start up for cfi-flash as well as environment and
state that depend on it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220108171356.586099-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>