soc: imx: gpcv2: split PGC domain probe in two passes
Currently it is possible that the platform device for a nested PGC
domain is added before the parent PGC device is there, which leads to
-EPROBE_DEFER when probing the driver. With normal probe this isn't
an issue, as the probe will be retried. With deep-probe this is fatal,
as the PGC domain devices aren't probed from DT, but via registration
of platform devices from the GPC driver, so the usual deep-probe
approach to ensure the devices are probed before the lookup isn't
working in this case.

Make sure to register the PGC domain platform devices in the correct
order to avoid the EPROBE_DEFER altogether.

Signed-off-by: Lucas Stach <dev@lynxeye.de>
Link: https://lore.barebox.org/20220116213221.3466936-2-dev@lynxeye.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>