Documentation: add chapter on Talks about barebox
There's been a number of talks about barebox. Collect them in a new
chapter for those, who prefer something less dry than wading through the
docs. A top-level chapter is chosen, so it's more visible to visitors.

The ordering is reverse chronological, except for the first entry: The
newest talk is German about running DOOM in barebox, which is a bit
detached to how barebox is normally used.

Instead, let's keep the first entry as a pointer to the latest overview
talk and then use reverse chronological order for the rest.

Signed-off-by: Johannes Schneider <JSchneider@jetter.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220124151827.3313412-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>