serial: cadence: add ZynqMP compatibles
Commit 3f2f5980d517 ("dts: update to v5.16-rc1") changes the compatible
of the ZynqMP uarts to "xlnx,zynqmp-uart" and drops the "xlnx,xuartps"
compatible.

The driver worked just fine before and the difference between the r1p8
and r1p12 compatibles is the use of the RX byte status register which is
only used with interrupts.

Add the "xlnx,zynqmp-uart" and the "cdns,uart-r1p12" compatibles to the
driver.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220202101119.3924600-1-m.tretter@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>